Create README.md
This project was made as a beginner project so that I am improve my python coding skills. I originally wanted to create a simple guestbook that allows
users post a message and other users to comment on the message. I am still learning how to work on this code and improve it. I am currently trying to turn
the code into a simple app on Heroku.